Removed redundant url key level from cache
Enough to cache by navigation event. Not sure why the url key was there - even more confusing now that the url is crumbless. Guess it was there because of how the code had evolved?
The history cache needs the crumbless url because it wants single scene for that url - wants to overwrite if revisit same crumbless url.
But the rscCache never wants to overwrite cache item - once react commits the tree then there's only that one item in the cache anyway. While things are in flight never want to overwrite  - it was already caching by navigationevent anyway just with redundant url key first?!