Supported parent CoordinatorLayout
The tabbar flickers on load when in edge-to-edge because the resize is async - React Native doesn't support sync resize yet. So the trick to removing the flicker is to do the resizing in Android instead of React Native. For this, need a parent with needsCustomLayoutForChildren so that React Native leaves it alone. CoordinatorLayout is obvious goto because BottomNavigationView should support being in CoordinatorLayout anyway (for example, HideBottomViewOnScrollBehavior).
Gave gravity bottom otherwise CoordinatoryLayout puts the tabs at the top. Then, once the insets are applied, trigger a sync measureAndLayout. Had to measure with UNSPECIFIED for the height otherwise the inset padding isn't included