Fixed Unregister scene in stack unit tests
Typical failing test was "Unregister current scene in stack Stack" where navigation A --> B --> C and then rerender and remove sceneC from the stack. That should navigate to A - so unfreeze A and freeze C. C stays in the stack for forward browser history - keep the scroll position etc. But A wasn't unfrozen and C wasn't frozen.
It started failing because the navigation is wrapped in startTransition - removing this fixes the tests. But can't remove it because want navigations wrapped in start/useTransitions so that they work with Suspsense, e.g., delay animation unit first Suspense boundary hit.
The problem is that C is 'unmounted' and so is instantly frozen - and infinitely frozen. So it's basically suspended forever. So React can't ever commit the changes - the state change callback in AsyncStateNavigator isn't called and nothing happens.
Fixed it by ensuring the suspension doesn't happen straight away - used the rest indicator to unfreeze. It was already there for the crumbs so used it for the unmounted too. But that isn't enough because if navigate back to A --> B then back again to A then scene C would become visible. That's because it's unfrozen and it was the freeze that was setting it's display to none. So manually set display to none when unmounted. This actually didn't work because the Scene didn't re-render - it's shouldComponentUpdate prevented it. So changed Scene to always re-render when the navigationEvent prop changes. That's enough to fix it.