Ensured crumb scenes frozen when duration 0
Tried to recreate the tests failure running the Twitter mobile sample. Could only recreate with duration 0 - think because navigating back from A --> B --> C to A --> B didn't 'unmount' C instantly because there's an animation. Then going back to A, React isn't bothered that C is frozen infinitely because not a change in freeze state maybe?!
So only duration 0 would recreate because navigating back to A --> B then C instantly frozen and React can't then commit changes.
Noticed that duration 0 didn't freeze A when going from A --> B. So fixed that by always animating. JavaScript happy with duration 0 animations anyway