Checked element exists before unobserving
Tested twitter sample with Home --> Tweet --> Tweet and removed Tweet scene from stack in a timer. After time up the Navigation router goes back to Home because Tweet invalid. Then when click Timeline it throws the error that the ref.current element is null. Not sure what makes it so, probably something to do with suspense.