Went back to jsx react and importing React
Originally changed because the typescript compile puts quite a few lines in the cjs to handle the import * and this is now in multiple times since not rolling up. But it breaks the app.html sample because React doesn't publish umd build for jsx-runtime (it's stopped publishing umd's for React).
Want to keep the changes to minimum and there's not much extra stuff in the files. There's nothing extra in the esm files anyway. Plus node is aligning them anyway so not worth changing anything for now.
Also support pre-React 18 where there's no startTransition - this is the app.html sample. Already checked this in LinkUtility so also checked in AsyncStateNavigator and NavigationHandler