Made RSCErrorBoundary parent of SceneRSCView
Otherwise it wouldn't catch the navigation 'error' in webpack sample. It did catch it in parcel sample but maybe threw it from Friends so the Person one caught it? Or maybe parcel reported it as coming from Person instead of SceneRSCView so the ErrorBoundary would catch it. Anyway, makes sense that it's this way round.