Ensured RSCErrorBoundary catches navigation
Can't have RSCErrorBoundary above SceneRSCView because then SceneRSCView disappears when the navigation 'error' happens (error boundary doesn't render children and it's now a child). If it disappears then it's not there to listen for subsequent navigations. Testing by navigate in Person if id is not 2 then if navigate to id 1 then get 'error' back to People then navigate to id 2 and scene is blank - because Person scene view no longer there.
So moved RSCErrorBoundary back down. But needed to create an extra nesting for the actual scene view otherwise the error originates in SceneRSCView and the boundary can't catch. The extra nesting means it catches it because the error originates in the nested component.
Also had to clear the error otherwise it stays in state in RSCErrorBoundary - ends up with a blank screen because NavigationEffect stays rendered. So cleared by setting error boundary key to the url so it remounts on navigation.