Prevented error boundary clearing scene state
Can't make the error boundary key the url because then scene state is cleared whenever do a refresh navigation, for example!! Was doing this so that the error is cleared so don't get blank scene when non-error navigation after an error. Testing person navigation 'error' when id != 2 then would get error but would then get blank screen when go to id 2 because there's still an error.
Worked out another way to clear the error - tied the error to a stateContext and then cleared the error whenever the stateContext changes. So next navigation clear the error. Then only go down the error scenario if the error context matches the current state context