Resolved server navigation links on the client
The server state navigator has the crumbless url - not the full url. Don't want the server to have the client context because, for example, want to be able to prefetch and that wouldn't work if prefetch depended on crumbs (would have to keep prefetching every navigation). So can't resolve the links on the server. For example, navigateBack would always error on the server because there aren't any crumbs.
So, instead, send the navigate instructions to the client and resolved them there were the state navigator has the full context