Removed server navigation
Although server navigation was working, not convinced about it. If tracking crumb trail, problem is that if navigate A --> B then on server navigate to C then the stack is A --> B --> C - on mobile, B and C would transition in together. Is that what's wanted? More likely want to ignore B so end up with A --> C, but can't do that on the server because stack isn't available on server.
Think only need to navigate on server if support RSC actions because need to redirect after save, for example. But that's easier done without RSC actions - standard fetch and then redirect on client.
Keeping error boundaries because can't really do error boundaries in userland. Can't wrap SceneView with fallback ui because need the SceneViews to always be there listening for next navigation. Fallback ui replaces them so no good. Has to be done by SceneView so it can do the replace as a child instead of as a parent