Removed redundant code
Tested without this and it works. The error is cleared by logic in getDerivedStateFromProps so don't need to check the contexts match in the render. Tested by throwing error in Person if id != 2 and then added RefreshLink on error fallback to go to Person 2 and it works.
Without getDerivedStateFromProps then the error is remembered and the navigating on error fallback still shows the error fallback