Removed server navigation from parcel sample
Already removed from core so this is the only bit that's left. Doesn't make sense to navigate on the server - only really relevant for rsc server actions where need to navigate after action complete. But easier to do with normal fetch then navigate on client. Also for ssr can 302 in middleware, for example