Made error boundary work for 'spa' scene views
Can't put an error boundary around a scene view because the fallback hides the scene view, stopping it from listening for navigation events. So it has to be a child and that's tricky to add in userland. So made the error boundary work for spa SceneViews like it does for RSC's. The more similar the api's are the better