Ensured declarative stack primed on first render
With declarative stack the state context isn't populated on first render. It's only populated by navigating firstLink in useEffect. It's only one missed render but it is noticeable if navigating from A --> B where B is TabBar with stack per tab. The tabs scene B is blank when it starts to transition. Only half way through does the tab content appear after that useEffect runs.
Also, if B has lots of tabs with heavy first scenes then the transition can stutter when the content is rendered half way through.
Workaround is to prime the stack(s) manually by navigating to first state when creating StateNavigator.
Changed stack to automatically do the priming (if not already primed with initial navigation). Stack creates a temp NavigationContext that is used for the first render - it's immediately replaced when the useEffect runs, but it's done it's job by then. Tested and it does get rid of the blank tabs scene - need to get @afkcodes to test whether stutter is gone, too, because he noticed it. Should be because when the useEffect runs, although there's a new navigation event, there's no render change so nothing to apply