Navigated link to populate peek context
It worked when priming the NavigationStack so applied same idea here. Better to navigate link than to manually build the context. Can't navigate from this context because it's not hooked up to the AsyncStateNavigator (and the real state navigator) but that's ok because this is really only a temporary context. As soon as the peek completes then a navigation happens and the real context is populated