Sketched in idea for mobile scene rsc
Same idea as on web except the rsc 'unit' is the scene on rather than the sceneview (maybe can support sceneview on mobile too for finer-grained updates). Simplified SceneRSC because shouldn't need history cache, for example - on mobile the scenes in history are still rendered in the stack. Also there's no ancestor check because it's the scene-level.