Got people --> person roughly working
The SceneView in Person component caused problems. Don't support SceneView yet so this was copied over by accident. But the bigger problem seems like some React bugs?!
Couldn't add and fetch the new scene at the same time (verified by only showed fetched rsc after setting state in useEffect in SceneRSC). Removing the wrapping div in Scene fixes this - so moved the wrapper into SceneRSC for now.
The useLayoutEffect in NavigationAnimation runs when before the children are committed so the scenes length is 2 but children is still 1. Added defensive check for now.
These problems might be because of the setState in the render (in both NavigationAnimation and NavigationStack) - don't know they because there's no problem if not fetching rsc's. Normal running then it all works fine?! Anyway, gonna try making the setState in NavigationAnimation a useMemo to see if that fixes the useLayout problem. (Did 'fix' it by changing it to useEffect and kept calling setState until the counts match - but obviously this isn't good)