Tried to get back navigation working
The scene children don't change when navigating back - because keep the person details in for history navigation. So updated an attribute so can observe that.
It kinda works but have to click person search back link twice. That's only the first time - after that can then click person details and person search back. But if change search, by sorting or paging for example, then need two clicks again.
Seems that the scenes.all isn't updated by the time the observer runs?! don't know what kind of bug this is with React where the state isn't in step with the committed dom?!