Fixed double click back to search from details
The mutation observer ran as soon as the dom committed the new count attribute and this was with the stale scenes state. Need the mutation observer to run after the mutation observer is added with the latest scenes. So did the actual move in a subsequent useEffect - the mutation observer just indicates the move needs to happen