Removed suspense otherwise user can't suspend
The implicit freeze/suspense added by NavigationStack for performance prevented the incoming scene from suspending - as soon as react hits a suspense boundary it commits to the dom. Tested by adding a 3000ms delay to the getPerson and the url updated immediately - because of the freeze.
Removed the freeze and the url updates after 3000ms when the data returns and the scene renders. So can't use freeze for rsc. Will have to wait for react.activity (old name offscreen) to add the same performance characteristic for scenes hidden in stack