Reused SceneRSCView in mobile
There's no need for a SceneRSC!! the one from web works.
Removed the clear rsc cache because this won't work on mobile - there are multiple scenes each with their own cache entry. Think it wasn't even safe to clear it on the web anyway - the useEffect could run while React is doing a background fetch.
Need React to say when it's safe to clear the cache - when React throws away the work without committing