Used SceneView from navigation-react
Removed SceneRSCView from the public api - it shouldn't have been there before mobile anyway and mobile doesn't need it anymore anyway.
Added a special __scene prop so the NavigationStack can tell which children are scenes with rsc. It can't use type because the type is the SceneRSCView. Can't add the _scene prop for non-rsc because the props come from user