Prevented double fetch on browser forward
Navigate People --> Person then navigation back to People then browser forward to Person. This would fetch the person scene twice. It was incorrectly cleared from cache when navigate back - the person useEffect cleared it but it shouldn't have run because it wasn't the current scene. Got away with this in web because there was only one scene around at a time - but on mobile there's many scenes in the stack and the person one isn't even in the stack but stays around for browser history. So added check to prevent the history stuff from running if not the current scene