Ensured data changed if crumbs length changed
Added dataKeyDeps={[]} sot person scene so that it never updates. Tested People --> Person --> Person (navigationlink instead of refreshlink when selecting a friend). This showed a blank scene instead of the friend. That's because dataChanged returned false instead of true - the old state and current state where the same so it returned false.
It was missing the check on the crumb length change. The data can only not be changed if it's refresh navigation. So if the old state is different OR the crumb length is different then the data has changed. So added crumb length check