Ensured history cache runs after history added
Load on Person scene then navigate back to People then browser forward and back and the People scene fetched twice. That's because there was no history in the cache. It wasn't cached because when navigating back to People the useEffect fired before the browser history was updated - so it failed the url check and didn't cache.
Added history event listener so that can run the history cache check after the history is added - need to run the history cache twice because the useEffect might fire after (normally does) browser history updated - so need to cache then. If it fires in other order the event listener will handle the cache