Replaced ImageSource with manually shaped type
Not sure why but ImageSource type doesn't work on react native 0.79. It compiles fine but the property isn't sent across to native. So replaced with a manual type that has the 4 properties used on native side. On ios turned that into ImageSource needed for loading. Not changes on android because it just extracts the properties anyway