Displayed images on non-primary tabbar
Applied image type changes already done for primary tabbar but the images didn't show up?! checked back on twitter sample and they don't even show up there so nothing to do with 0.79 or the type changes.
Turns out its the multiple calls to setupWithViewPager - only need 1 call and calls after that clear the images (and maybe other things?!). So added check to bypass if alreayd setup with that viewpager