Cleared rsc cache
When the navigation event changes then safe to clear the cache of the old navigation event. The old navigation event is the old one for that rsc scene view - they only have one cached value in use so safe to delete. This idea works on react native too - scene B won't clear out scene A because B and A have completely different navigation events.
Also cleared out uncommitted scenes from the cache. For example, added a delay of 300ms to people search and filtered by 'Adit' - this fires off 4 state updates. React only commits the last one when it returns and abandons the other 3. So they're not needed in the cache anymore. Stored a __committed flag against the cache and then cleared out anything uncommitted when something is committed.