Sized navigation bar when there's overlap
On new architecture the touchables did nothing on the timeline scene on physical android device. On physical device the onTouchMove is called (not called on simulator) and the js side cancelled the event because the move was deemed outside the responder region. Obviously it wasn't - but it's because fabric measures based on the shadow tree not based on the host tree (think that's right). And the shadow tree didn't have the correct position for the scrollview (and hence its touchables). So included the overlap in the calculation.
Basically, on fabric need to make sure that React Native has the correct position for the navigation bar. Get a similar problem when the toolbar collapses and the touchables don't do anything - well they do if press at the bottom. But pressing at the top does nothing because again React Native thinks its outside responder region. Going to have to allow for the offset in the height