Sized navigation bar when for offset
Same reason as before - on new arch the pressables don't work if the navigation bar isn't size correctly. React Native cancels the touch onTouchMove (fires on physical device) because it thinks its outside the responder region. Sizing the navigation bar fixes this. So took into account the offset when scrolling a toolbar - also took a fraction of the overlap based on the offset (got the idea from getOverlapRatioForOffset in ScrollingViewBehavior). Had to delay the resize otherwise the collapsing header on timeline was jerky - waited for the scroll to finish before doing the resize