Copied changes over to the old architecture
This threw errors because of NaN values for height. So protected against that.
While doing that noticed that overlapRatio was 0 when half collapsed. That was because casting to int turned 0.8 into 0, for example. So moved the cast until after multiplying by overlap instead of on the ration part (will always give 0?!)