Restructured the rsc sample setup
Started documenting RSC and realised the setup was confusing. Didn't look the RootProvider. Plus there's no need to create a component for users to put their providers into - that's Next layout nastiness. They can put them straight into App if they want, just like in the old days. So created separate HmrProvider and NavigationProvider and inlined them into App.
Created a BundlerContext that has the setRoot and deserialize - used to be called fetchRSC but changed to deserialize after reading https://overreacted.io/why-does-rsc-integrate-with-a-bundler/
Simpler structure to use and document