Moved rsc menu to the end
Makes it clearer that all the docs apply to non-rsc and rsc alike. Also makes it easier to explain how to pass shared elements function props to NavigationStack in rsc-world. Prefer to do this in rsc doc than in shared element doc. Don't want rsc to appear in any other docs for clarity