Calculated dataChanged only once
Scenario that went wrong - was testing out useRefetch so created a FriendsClient in rsc sample and added
useRefetch(({data, oldData}) => !friends || data.show === oldData.show);
Idea is that only need to update the very first time and then never again if toggling show on and off - can do that all on the client. If changing id or gender then need to udpate (data.show === oldData.show).
But when load the person scene with show false and then toggle show on/off/on/off then friends component would disappear!
It was because it set the wrong renderedSceneView.current in useEffect - the first friends expanded the dataChanged starts off false but then in the useEffect it becomes true because friends is set by then.
So evaluated it once so the values is unchanged for that async render