Removed refetch on history navigation
On a master/details scene don't want the master list to refetch on history navigation because then it could be out of step with the details - the selected details might not be in the top 10 anymore, for example. Also, common to programmatically pop browser history when closing modal - would be bad if the master list was different.
Safest not to refetch on history navigation. Also better for mobile where there's no need to refetch - the crumb scenes are already rendered in the stack.