Sketched in untested idea for manual refetch
cons refetch = useRefetch();
refetch() - triggers manual refetch of the SceneView
useOptimistic to implement - make ancestorFetching false and ignoreCache true to for the rsc refetch. Cleared the rscCache as well - won't happen automatically because not mutating the NavigationEvent in context.
Flipped around the sceneView function so that it returns from cache if there's something in cache - because, once the refetch returns, the optimistic values are set to the 'correct' ones and so ancestorFetching might be true. But then it would incorrectly return children when really needs to return the cached values from the manual refetch.
Also plan for refetch(true) - triggers manual refetch of the Scene - can use the mutation of NavigationEvent with ignoreCache for that.