Supported vite hmr rerender with old url
If edit the state navigator route config twice in a row then the 2nd time vite rerenders the NavigationProvider from scratch with the new url bug the old stateNavigator - only temporarily, it then renders it with the new stateNavigator. So caught the error and prevented SceneRSCView throwing when there's empty state context.