Replaced deprecated SafeAreaView
Had to use 2 SafeAreaProviders otherwise Notifications didn't show properly. With only 1 in Notifications the tabs were hidden under NavigationBar and with only 1 at root the title on NavigationBar was hidden (think that's what happened). Also couldn't render the SafeAreaView on Android because it made the screen transition janky. Tried 'turning it off' on Android using the edges prop but was still janky.