Started transition only if not refresh
For refresh plan to useDeferredValue for navigation context so that some SceneViews and use the deferred and some can use the non-deferred. Hopefully will allow navigation data to control input fields for filtering, for example - they need to update immediately, not deferred. Only deferred value allows this control - once start a transition then everything is within that.