Sketched in non/deferred switch idea
If the SceneRSCView doesn't refetch - doesn't need to go back to the server - then use the non-deferred navigation context. If it does need to go back then use the deferred.
The scenario in mind is a list with filters. Want the filters to be controlled by navigation data and so they must update immediately. The list must be deferred because it needs to fetch results from the server. That's the classic deferred scenario in react docs but using navigation data to control the filter.
For now just checking if the refetch is empty array but plan to flesh this out so it checks if refetching - although can't think of scenario where only sometimes refetch?