Changed to use client scene view as context switch
Tried wrapping Filter in SceneView and realised that doesn't work - there's no data to fetch so don't want to make it a server-entry with a mapped component. Better to wrap the contents of the Filter in SceneView and then it's the client one instead of the server one - then made this the switch. So client scene views use the current context and server/rsc ones use the deferred context. That still works for non-refresh navigation because then current and deferred are the same anyway