Fixed recursive setstate error
Not sure why but React said there was recursive setState - went back to codesandbox (https://codesandbox.io/p/devbox/react-dev-forked-xlf9fw) to try and replicate the error. Couldn't replicate but didn't work until added the deferred into the useMemo - guess it kinda makes sense because the current doesn't change, the deferred eventually becomes the current.
Still not properly working but the error's gone. It seems to work until stop typing and React applies the deferred. Then it goes back to not deferred, blocking on typing - think it's because using the navigation event as the cache key and that's the deferred and then deferred becomes current and there's no key match anymore. The codesandbox doesn't have problem so working on that assumption (https://codesandbox.io/p/devbox/react-dev-forked-xlf9fw)