Prevented block after type - wait - type
Typed 'ad', for example, then paused to wait for the response to come back (list filtered to 'Aditya Larson') then typed again and this would block - the input wouldn't update as typed. The navigation event was cleared from the cache so it refetched for the old - think that React does a rerender for the old navigation context when doing deferred. That's been deleted so it goes into refetch and this blocks everything because it's not deferred anymore.
Need to investigate what's happening - but deleting this keeps the input typing responsive