Prevented block after type - wait - type
First type 'ad' the wait - then type 'it' and it blocked.
When typing single 'i' character (after the wait) then React renders with new context for the filters and old for the list. In the background React renders list with new context. When the set state for new with old resolves then the useEffect for filters (People scene) runs and deletes the old context from the cache. Then when type 't' the old context isn't there so the immediate render with new and old is blocking - because the old context isn't there it does a fetch which blocks.
Changed to only delete once the deferred is resolved.
Checked the rsc cache and there's too many things in - every character typed creates cache entry and they're not getting cleared. Need to investigate