Stored only deferred in new rsc context
Don't need both. If the scene rsc view isn't fetching then it can use the current navigation event in context. If this is the non-deferred then that's ok because want the latest/current if not fetching. If it's the deferred then can't change it back to current anyway - must use the same navigation event for all descendants