Started converting NavigationHandler to function
Don't like the NavigationHandlerInner and should be able to get rid of it when NavigationHandler is a function. Also should help with setting the history (resumeNavigation) when the transition/deferred has settled.
Moving the AsyncStateNavigator into the function component should also help with this because it has access to all the stuff in closure.
Lots of errors but want to commit because it's an ok start