Fixed tests by not storing function in state
The resumeNavigation function can't be put in useState because React executes it - thinking it's returning new state based on previous value. So remodelled and stored resume on the navigationEvent state directly - they go together anyway.
Now the only tests failing are the 3 'change state navigator' ones which are expected as haven't implement this yet