Made gender navigation data controlled
Same idea as already implemented for filter on people - need the person scene view to receive the current (non-deferred) navigation event and the friends will get the deferred one. That means the filter will update immediately so no need to useOptimistic. But the person has a refetch so implemented the proper fetch check - where checks if fetching and, if not, returns the non-deferred.
There was a tricky scenario
1. show friends
2. filter to female
3. select a friend
4. filter to male
This would delay the dropdown selection. It would make 2 fetches, first it would refetch female and then male - this caused the delay because the gender selection wouldn't update till the female was returned.
It was wrongly refetching because there was no cached scene view for female friends - when the friend was selected the parent person scene view refetched so the friends didn't and so there was no cached scene view. Then when changing to male React rerenders with the current navigation event for person and old navigation event for friends - so the person ancestor isn't fetching because only gender changed, but there's no cached scene view for friends. So it refetched. But in this case it should just use the renderedSceneView.current because it's render already committed for that old navigation event. So added a check for committed and returned false from fetching