Fixed history cache
Now that resumeNavigation (which writes to history) is called in different place the timing changed. It meant the history cache wasn't populated because the url wasn't written by the time the scene view useEffect ran. But the only [reason it needed to be added](https://github.com/grahammendick/navigation/commit/278507c71c63088afe6ca7959e262c7a66a1005b) was to stop the non-top scene deleting history in mobile stack.
Moving the caching into the NavigationHandler fixes this because it operates at the top level stateNavigator. So don't need to wait for the browser history to be written because its stateNavigator always relates to the top level scene.
Also don't need a global cache - needed that for the fetch results because can't useRefs for these background renders. But the history caching only runs in useEffect after the React tree is committed so can useRef for it.
So split the history cache logic into 2. Managing the scene count and deleting from the cache is in the NavigationHandler. And the scene views are in charge of caching their own rendered views.
Also means don't [need the popState listener](https://github.com/grahammendick/navigation/commit/37666f4d9a6a0e231e0e882e44bfd1610b839c7b) because that was legacy from when needed the url to be written.