Corrected the mobile rsc navigation
Navigate A --> B then screen A would go blank. It was using the wrong navigation event because the deferred was the top level one. So fixed the deferred navigation event same way as the navigation event is already fixed - one per scene - only changes when that scene is at the top