Fixed the mobile rsc refresh navigation
Navigation to person and click show friends but nothing happened. Clicking a second time did show friends.
When deferred the scene renders the first time with the old navigation event for deferred and then the second time in the background with the new navigation event for deferred. But the scene didn't update because the navigation event doesn't change the second time - only the deferred one changes. So changed shouldComponentUpdate to check if the deferred one has changed and to render in that case