Fixed refetch when history
Added test button to Filter on People that does an update of a person's name followed by refetch(). Navigate People → Person then browser back and press test button. It should show latest data but it doesn't. It works ok if navigate back to People but not if history. Also works ok if refetch(true) - the whole scene.
The problem is that the navigation context on the List scene view wasn't the refetch event. The List uses the deferred event and that was coming from the Navigation Handler which doesn't have the refetch event. The refetch event is held by the people scene view - it was overriding the navigation event with it but not the deferred event and the list uses the deferred event.
The problem only happened on history because the the sceneView returned is different - if not history it returns children which has the latest data because person refetched, but if history it returns the one from history cache which is stale.
Changed the scene view to override the deferred event with the refetch event