Supported ios26 full screen swipe to pop
Followed same pattern with the new interactiveContentPopGestureRecognizer as already in place for interactivePopGestureRecognizer - the latter is for the swipe from edge and the former is the full screen swipe.
Also added new UIPanGestureRecognizer following same pattern as already in place for UIScreenEdgePanGestureRecognizer - this is for custom animations. Can't use the built-in gesture recognizers when handling custom animations so need own ones. The screen edge was already there and the new one is for full screen swipe on custom animations.
The navigation router already supports the ios26 interactive while navigating - for example, can click back multiple times and it will go back each click rather than waiting for the transition to complete like pre-ios26. But couldn't see how to get this to work for custom animations - there's no guidances around it and suspect it only works for the built-in transition.
Tested the fullscreen swipe to go back when there's a [react native panresponder](https://reactnative.dev/docs/panresponder) on the page. They both run together unfortunately (both the in-built and custom animations) - guess because the react native one isn't really a PanGestureRecognizer so can't take precedence. Not worried about this - maybe the one from rngh works by handling the gesture by itself, but haven't tested.