Supported barbutton as shared element
In ios26 uibarbuttonitem are valid fluid transition sources. Tricky bit is getting at the bar button items from the sheet - implemented the NVSharedElement protocol and added all NVBarButtons to the sharedElements array.
Hard-coded the shared element name for quick testing and it works - added a sheet to the fabric home page and opened it from bar button item - chose eraser as name because used the eraser shared symbol for the bar button