Added delegate based on custom animation prop
Only want the delegate with animationControllerForOperation if custom animation otherwise can't fallback to default ios pop gestures - and end up with conflict with horizontal scroll. So added basic controller delegate if custom animation false and the transition one otherwise. Also only added the custom gesture recognizer if custom animation, too - the swipe from edge and full swipe - don't need them if not custom because ios handles it. And ios prevents the gesture conflict too - that's whole point of this work