Registered root scene views
Need the root ones to pass to the server. The server needs to know what roots to render when redirecting - if rsc render request is for scene A but redirect to scene B then server needs to get all the root scene views that are active for scene B. Then it will issue the render request for these.
Up to know the client has worked out what to render based on active and data changed but for redirect the server needs to do that calculation. It can't use data changed because then it would need to know about all the refetch calcs of all scene views - not possible to pass refetch calcs to the server because useRefetch means they aren't all serializable. But it's ok to render the root scene views in a redirect - so only need to know about the active props, which are all serializable