Used full link - including crumbs
In order to navigate on the server need to send the full link down. The plan is to navigate against the state navigator on the server before rendering rsc. Then render rsc using this latest state navigator - only way to handle a redirect without requiring a second render triggered from the client. Do it all navigation on the server and then rendering once. But need the full state context so can navigate against it.
It's better anyway because then server is aware of full context and could render something different if there's a previous scene to if there isn't. This means that history cache is on the full url - but that's better.