Sketched in client part of the server navigation
Detect if server navigated and then cache the response - use the navigate event data as the key so only get a hit for that navigation. Then when the navigation commits - in useEffect - check if server navigation happened and, if so, don't resume because don't want to update the url (or the real state navigator). Instead navigate to the new url returned from the server - cache the already fetched next scene view in a ref and then return this ref from the deserialize when called. It's just a rough idea of how it could work but there are issues
1. Get a blank screen until the rsc returns (added 1s delay for peron fetch) - the first promise resolves instantly to null so React shows blank screen (no rsc content). When the 1s is up the fetch resolves and shows content
2. Can't guarantee that the root scene view calls deserialize - if it's a refresh server navigation then might be only the sub scene views that call deserialize - but need it to be the root one